+++ src/components/VideoTile.tsx
@@ -1,18 +1,10 @@
 import React from 'react'
 
 import { useVideo } from '@/api/hooks'
-import { AssetAvailability } from '@/api/queries'
 import { absoluteRoutes } from '@/config/routes'
 import { AssetType, useAsset } from '@/providers/assets'
-import { singleDraftSelector, useDraftStore } from '@/providers/drafts'
-import { useUploadsStore } from '@/providers/uploadsManager'
-import {
-  VideoTileBase,
-  VideoTileBaseMetaProps,
-  VideoTileBaseProps,
-  VideoTilePublisherProps,
-} from '@/shared/components/VideoTileBase'
-import { copyToClipboard, openInNewTab } from '@/utils/browser'
+import { VideoTileBase, VideoTileBaseMetaProps, VideoTileBaseProps } from '@/shared/components/VideoTileBase'
+import { copyToClipboard } from '@/utils/browser'
 import { SentryLogger } from '@/utils/logs'
 
 export type VideoTileProps = {
@@ -50,59 +42,12 @@ export const VideoTile: React.FC<VideoTileProps> = ({ id, onNotFound, ...metaPro
   )
 }
 
-export type VideoTileWPublisherProps = VideoTileProps &
-  Omit<VideoTilePublisherProps, 'publisherMode' | 'videoPublishState'>
-export const VideoTilePublisher: React.FC<VideoTileWPublisherProps> = ({ id, isDraft, onNotFound, ...metaProps }) => {
-  const {
-    video,
-    loading,
-    videoHref,
-    thumbnailPhotoUrl,
-    avatarPhotoUrl,
-    isLoadingThumbnail,
-    isLoadingAvatar,
-    uploadsStatus,
-  } = useVideoSharedLogic({
-    id,
-    isDraft,
-    onNotFound,
-  })
-
-  const draft = useDraftStore(singleDraftSelector(id ?? ''))
-
-  const hasThumbnailUploadFailed = video?.thumbnailPhotoAvailability === AssetAvailability.Pending
-
-  return (
-    <VideoTileBase
-      uploadStatus={uploadsStatus}
-      isLoadingThumbnail={isLoadingThumbnail}
-      isLoadingAvatar={isLoadingAvatar}
-      publisherMode
-      title={isDraft ? draft?.title : video?.title}
-      channelTitle={video?.channel.title}
-      channelAvatarUrl={avatarPhotoUrl}
-      createdAt={isDraft ? new Date(draft?.updatedAt ?? '') : video?.createdAt}
-      duration={video?.duration}
-      views={video?.views}
-      thumbnailUrl={thumbnailPhotoUrl}
-      hasThumbnailUploadFailed={hasThumbnailUploadFailed}
-      channelHref={id ? absoluteRoutes.viewer.channel(video?.channel.id) : undefined}
-      isLoading={loading}
-      onOpenInTabClick={isDraft || !id ? undefined : () => openInNewTab(absoluteRoutes.viewer.video(id), true)}
-      onCopyVideoURLClick={isDraft ? undefined : () => copyToClipboard(videoHref ? location.origin + videoHref : '')}
-      videoPublishState={video?.isPublic || video?.isPublic === undefined ? 'default' : 'unlisted'}
-      isDraft={isDraft}
-      {...metaProps}
-    />
-  )
-}
-
 type UseVideoSharedLogicOpts = {
   id?: string
   isDraft?: boolean
   onNotFound?: () => void
 }
-const useVideoSharedLogic = ({ id, isDraft, onNotFound }: UseVideoSharedLogicOpts) => {
+export const useVideoSharedLogic = ({ id, isDraft, onNotFound }: UseVideoSharedLogicOpts) => {
   const { video, loading } = useVideo(id ?? '', {
     skip: !id || isDraft,
     onCompleted: (data) => !data && onNotFound?.(),
@@ -117,10 +62,6 @@ const useVideoSharedLogic = ({ id, isDraft, onNotFound }: UseVideoSharedLogicOpt
     assetType: AssetType.AVATAR,
   })
 
-  const uploadsStatus = useUploadsStore(
-    (state) => state.uploadsStatus[video?.mediaDataObject?.joystreamContentId || '']
-  )
-
   const internalIsLoadingState = loading || !id
   const videoHref = id ? absoluteRoutes.viewer.video(id) : undefined
   return {
@@ -131,6 +72,5 @@ const useVideoSharedLogic = ({ id, isDraft, onNotFound }: UseVideoSharedLogicOpt
     thumbnailPhotoUrl,
     avatarPhotoUrl,
     videoHref,
-    uploadsStatus,
   }
 }
