+++ src/types/uploads.ts
@@ -1,0 +1,6 @@
+export type AssetUploadStatus = 'completed' | 'inProgress' | 'error' | 'reconnecting' | 'processing'
+export type UploadStatus = {
+  lastStatus?: AssetUploadStatus
+  progress?: number
+  retries?: number
+}
